Media is now an engine. There's always an instance at /admin/media (it's in the fixtures and in the data migration task). You can make more instances. You can categorize media. You can lock media engine pages down to particular categories.
YOU NEED TO RUN ./symfony apostrophe:migrate-data-from-pkcontextcms --env=appropriateenvironmentname

After updating to this latest version. You will get warnings about things that can't be done because they
were already done on previous runs of the task. That's OK.

You do NOT need to run the code migration task, only the data migration task.

Note that the media module will move to /admin/media as a result of this change.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@175 c89fab08-2f9b-42bf-8176-bd3ad6bee547