When netpbm is disabled in favor of gd (app_imageconverter_netpbm set to false or netpbm not installed), Rendered PNGs now have alpha channels and preserve the alpha channels of PNGs that are uploaded.
TODO: I need to fix the persistent file upload previewer so that this doesn't seem broken when you first upload the image (due to the use of JPEG for all previews, which was a lazy hack on my part). Once you hit save in the media repository you'll see that your alpha channel is actually intact.

TODO: newer versions of netpbm can allegedly handle alpha channels too. See if that's true in practice and how likely it is to break a lot of existing installs. There are new utilities for this purpose apparently.

Note that if you try to force JPEG or GIF format on that alpha PNG it'll composite against black. This is gross, but our image slots never do that. Anyway, if you're going to upload something as PNG and render it as JPEG, flatten it before you upload it.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@430 c89fab08-2f9b-42bf-8176-bd3ad6bee547