Apostrophe functional test support merged. Should work for both the sandbox project and various client projects that have a different prefix for URLs (see setOptions in the functional test base class). I'll need to merge those further improvements back down to stable.
Merged in support for fast reloading of fixtures and correct loading of fixtures from plugins etc. 

aNavigation has some static caches, so it needs to respond to the simulateNewRequest event. Otherwise things are unexpectedly already populated when we move on to the next functional test, with inconsistent results. To make this simple I call a static method there from aTools::listenToSimulateNewRequestEvent(). We can add more calls in that method as needed as long as they are for things in apostrophePlugin, code in the blog plugin or similar will need to implement listening to that event in its own right.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@755 c89fab08-2f9b-42bf-8176-bd3ad6bee547