Pretty signficant changes to the history browser.
The history browser now has its own close button instead of piggy-backing on the area cancel button.

The area cancel button has now gone away completely because we use the nifty dropdown for Add Slot

And History has it's own set of controls.

Additionally, The Save as Revision and Cancel Preview buttons live in the Preview container.

The preview container received new styling to be position fixed.

TODO: Test in IE7 before the 1.1 release of Apostrophe.

I cleaned out the majority of the cruft in aUI.js so that it contains only the esesntials and none of the weird old stuff that it had before

Also cleaned up some of the singleton slot controls js. We are still unfortunately using jQ to yank up the singleton slot controls into the area controls UL.






git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@1040 c89fab08-2f9b-42bf-8176-bd3ad6bee547