New aImageConverter::supportsInput($extension) method allows you to check whether aImageConverter can import a particular image format on this system. Mainly used to check for pdf support:
if (aImageConverter::supportsInput('pdf')) { ... }

You can pass:

gif, jpg, jpeg, png, bmp, ico, pdf (pdf always reports false if ghostscript is not available) and, if you have the netpbm utilities, other well known file extensions supported by netpbm.

This method can of course be fooled by broken installations in which the netpbm utilities or ghostscript are present but not working. 

Lack of support for bmp and ico is neither a bug nor uncommon.

For performance reasons the result is cached for 5 minutes. Keep that in mind if you make system changes such as installing netpbm and ghostscript and see no immediate difference.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.3@1077 c89fab08-2f9b-42bf-8176-bd3ad6bee547