No more hardcoded routing.yml in apostrophePlugin to break people's backend apps. Instead we have a second event which registers when app_a_admin_routes_register is set (that defaults to true), and registers routes for the various admin modules only if they are actually enabled. Much more conservative.
Fixtures for the /admin page now correctly specify the aAdmin engine (this was a bug, previously unreported). The route for the aAdmin engine has the correct class. Now users are correctly redirected to the @homepage if they try to access /admin directly (we might do more with that URL later).



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@1391 c89fab08-2f9b-42bf-8176-bd3ad6bee547