Hooray, you can link directly to an engine page just by passing an engine-slug parameter as one of your route parameters in link_to and url_for! No more aRouteTools::pushTargetEnginePage() (although that is still occasionally useful and fully supported).
When I tried to do this with Symfony 1.2 I couldn't make it work because 1.2 always did in-memory caching of routes by their parameters for generating routes in the current page, even if you disabled the routing cache. In 1.4 routing cache off means routing cache off (and Apostrophe already requires it be off), so we're good to go.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@1408 c89fab08-2f9b-42bf-8176-bd3ad6bee547