Improved a_js_call() syntax with '?' placeholders allows full flexibility in how you call js methods, including the use of parameters on the left hand side, as in:
myobject.property['body'].brighten(50);

Which can be written as:

<?php a_js_call('myobject.property[?].brighten(?)', 'body', 50) ?>

If you really need to pass in your parameters from somewhere else, you can use a_js_call_array:

<?php a_js_call_array('myobject.property[?].brighten(?)', array('body', 50)) ?>

If you need it you'll know why. (:

Thanks to Stéphane and others on the apostrophenow list for their contributions to the discussion of this feature.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@1975 c89fab08-2f9b-42bf-8176-bd3ad6bee547