Re-worked the page settings form significantly. There's some new markup in the form, some new CSS as well.
This work is incomplete, there still needs to be CSS cleanup across the board for Create page form and Page settings form, there needs to be some Javascript added to the forms to highlight the focused section as well.

But we're getting there and it's look good.

Some more CSS clean up in general as well



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2135 c89fab08-2f9b-42bf-8176-bd3ad6bee547