Resolves #436
Something for the I18N folks:

You can now put :sf_culture in the a_page route if you wish, like so:

a_page:
  url:   /:sf_culture/:slug
  param: { module: a, action: show }
  requirements: { slug: .* }

This allows Google to index separate pages for separate languages and allows links to language-specific pages to be shared by users, etc.

This leads to the question, "how does Google ever see the home page in a language other than English?" Our default language switcher just changes the user's culture and redirects to the home page without a culture in the URL. But you can also access culture-specific homepages at:

/de/
/en/
/fr/

Etc.

Consider removing the standard "homepage" route and replacing it with an action that redirects to the home page URL for the default language.

Note that we have not attempted to internationalize slugs.

Thanks to Niko Viitala (sorry Niko, I didn't at first realize from your ticket that it was such a simple change or I would have done this sooner).



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2496 c89fab08-2f9b-42bf-8176-bd3ad6bee547