aMultipleSelect now has an "autocomplete" option. When autocomplete is set to a URL, the user sees a text field instead of a select menu and the options they select via jQuery autocomplete are added to their selection. Consider:
$this->setWidget('blog_posts', new sfWidgetFormChoice(array('multiple' => true, 'expanded' => false, 'choices' => array())));
$this->validatorSchema['blog_posts'] = new sfValidatorDoctrineChoice(array('model' => 'aBlogItem', 'multiple' => true, 'required' => false));

...

aMultipleSelect('#a-<?php echo $form->getName() ?> .blog-posts', { 'autocomplete': <?php echo json_encode(url_for("aBlogAdmin/search")) ?> });

This is very reusable. It's currently suitable only where users are not allowed to add entirely new items - the only way to add an item to the active selection list is by selecting it from autocomplete.

Extremely handy when there are too many items for a list but a list is conceptually correct (for instance, blog slot post selection).

It needs styling.

Part of the resolution of #623



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2529 c89fab08-2f9b-42bf-8176-bd3ad6bee547