* Refactored my update tab functionality so you can use it anytime. Call apostrophe.update('#mything'), then later if you want it to change from 'updating' to 'updated' and disappear on its own, call $('#mything').trigger('aUpdated');
It's still automatically called if you use apostrophe.formUpdates().

I just implemented this for the "new post" title dialog and its "save" behavior is a lot less baffling now although I could stand to work on it more.

John and Rick... if you change the visual behavior of this please think about how to keep it general and awesome and help the dev team to make more intuitive ajax interfaces without angsting over each and every submit button's "I am updating" behavior. Take advantage of what I've done here with jQuery's offset method to superimpose a better spinner somewhere if you like, but think about how to locate it in an automatic and natural way, with the reasonable fallback behavior I already have here or similar. Maybe you'll extend my "find a better ancestor for the tab to live on" feature in the other direction and look for a descendant element where a spinner should live (ie spinner floating on submit button) for instance.

Bottom line... if someone just makes a form and calls apostrophe.formUpdates() on it to make it ajax, something halfway reasonable now happens from the end user's perspective, and that should continue to be the case.

* Quieted down warnings when engine pages don't exist - throwing an exception is plenty loud enough.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2598 c89fab08-2f9b-42bf-8176-bd3ad6bee547