YOU MUST apostrophe:migrate AND THEN rebuild-search-index AFTER THIS UPDATE OR YOU WILL GET 500 ERRORS ON SEARCHY THINGS
Implemented #684. Most of the time went into figuring out a way for the start and end times of the event to show up, even in sitewide search, without pulling all the event objects. Not showing start and end times for events in search results seemed too lame to be worth displaying.

The new solution is a neat one I think: let's say you have a virtual page that's associated with an object of the aEvent class. Set the engine field of the virtual page to aEvent. When the virtual page is indexed by the search engine, it will look for an aEventSearchHelper class, instantiate it, and call filterUpdateLuceneIndex() with the options about to be passed to aZendSearch as the argument. Add more stuff and return the arguments. Now you have start_date, end_date, etc. in the search index.

Your aEventSearchHelper class must also implement a getPartial method. If this method doesn't return null, then the partial it names is rendered instead of the default rendering for the search results for a page. And the Lucene result object passed to it has the extra stuff you stored in the filter method. I made sure those fields were compatible with the aEvent/dateRange partial... boom done

It's possible to add the location field to this scheme if we feel it's worthwhile to display it in the search results. It would be smart to do it sooner rather than later so one rebuild of your search index will suffice.

The aEvent/searchResult partial can be styled as needed.

Note that getFirstEngine() and the blog plugin's best-engine-finding methods have been modified to carefully avoid returning a virtual page. That's something to take care with now that virtual pages have a value in the engine column.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@2936 c89fab08-2f9b-42bf-8176-bd3ad6bee547