Fixed #448: a blog posts slot with a post in the jobs category will now link to the jobs-specific engine page.
The aDoctrineRoute class now automatically calls getEngineSlug on the sf_subject (the object the link is for), if that method exists. So far in my tests this looks like a solid fix to the longstanding problem of generating decent permalinks for blog and event slots.

This is done only if we are not already on an engine page of the relevant type (for instance you are never taken away from the /blog engine page when you click on a permalink for something shown there, even if it's even better for the /job page). 

Also you are never taken away from an engine page that has been explicitly pushed or set with the engine-slug parameter.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@2964 c89fab08-2f9b-42bf-8176-bd3ad6bee547