Replace <p>&nbsp;</p> (generated by all the major browser RTEs, or possibly by FCK itself) with <br />. Implements #500.
There are other cases, like <h4>&nbsp;</h4> if you put a "blank line" between h4's, but I don't think we should change these because the visual impact of a br between two separte major heading elements is not the same as a "&nbsp; only" major heading element in between them. A deeper fix would be something like:

<h4>First line<br />Second line</h4>

But this change is not trivial to apply because we have to remove the closing and reopening of the neighboring h4's to merge them into one, and the styling of the site has to be written with the possibility of internal br's in headings in mind (often folks assume br won't appear in headings at all and give it a fixed height). More discussion is welcome.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@2975 c89fab08-2f9b-42bf-8176-bd3ad6bee547