_mediaHelper: wrapped mediaClearSelectingOnNavAway in an IF isSelecting()
a.js:	added href === undefined check to mediaClearSelectingOnNavAway()

audio:	added href="#" to play and pause button

services logos css updated to work without a logo if it is not styled. A simple txt label is displayed instead

added a-ui class to all div.a-media-select. Some had it and some didn't. This was resulting in inconsistent layouts of this ui element




git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3037 c89fab08-2f9b-42bf-8176-bd3ad6bee547