Implemented [847]
I screwed my courage to the sticking point and hit this one properly.

You will now see only links to pages you are allowed to visit, period. The logic should be as good as that which applies when you actually try to visit the page. But the queries should still be swift.

I also embarked on a refactoring of the getChildrenInfo, getAncestorsInfo, etc. methods. These are now wrappers around better implementations at the page table level which take an options array. In most cases the livingOnly option is now simply ignored for bc purposes in favor of showing you only what the user is permitted to see. Exceptions are methods where some higher level filtering really is necessary, as with the tree nav component. Unfortunately there is not time to refactor every last one of these methods, or to introduce options arrays to the methods at the page class level, but this is still a big improvement.

This was a big change. I did a lot of testing but more testing would be very helpful. Let us know your results.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3099 c89fab08-2f9b-42bf-8176-bd3ad6bee547