Updated error404Success and noMediaItemsArea a bit.
Moved a-footer out of a-wrapper so we could have a footer that spanned the width of the browser window.

Closed ticket #137 - Areas with one slot type: default to "add slot type" and not display the area menu. Look at _area.php and _addSlot.php. I just had to shuffle a few things around and this fell right into place.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3130 c89fab08-2f9b-42bf-8176-bd3ad6bee547