Fixes [956]. All slots that use aImageSlot/choose to pick a media item now automatically get the fix. I also fixed aSlideshowSlot/normalView, if you have written code that needs to pick a slideshow rather than one media item you'll need to adapt the fix for your use.
The fix is very simple:

* Stop setting actual_url or actual_slug in 'after'
* Set the a-inject-actual-url class on the a tag.

a-inject-actual-url triggers new code in apostrophe.smartCSS that locates 'after' and adds 'actual_url' to it correctly based on window.location.href, which is always the actual page and not some AJAX action or other (which is what confused the pure PHP solution).

I also added apostrophe.parseUrl() and apostrophe.decodeQuery() which are useful for constructing solutions like this.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3454 c89fab08-2f9b-42bf-8176-bd3ad6bee547