Resolves #1021, #1022, #1023: all pages tested are 10%-25% faster when app_a_fasthydrate is set to true. I'm doing manual Doctrine hydration of the critical things (the page and the slots in the slot cache) and leaving the intermediate data structures in array form.
Also: phpdoc for the core API methods. Renamed getArea() to getSlotsByAreaName (there is a bc wrapper for getArea() calls, so they will still work). Direct access to previously memory-cached pages in aTools so you can leverage that functionality outside of an a_slot call if you need to.

Since I'm changing something pretty important (hydration) I made this a feature flag that defaults to false at the code level, however it is switched on in the sandbox so we should get feedback on it and it will be in our new projects. We can enable it for dev in old projects and test safely before deploying the new feature to production



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3552 c89fab08-2f9b-42bf-8176-bd3ad6bee547