Fix for #931. Jeremy had a simple and effective idea (also voiced by others) to use the existing capabilities of the minifier to group related CSS and JS files. We add a data-asset-group attribute to each link and script tag; the minifier is already smart enough to know that things that don't have the same attributes can't be minified together, so it automatically groups accordingly. Now we are no longer recompiling the same assets just to add one file that happens to be used by a specific action.
We've added an app_a_asset_group option in app.yml which specifies the asset group name for apostrophe's core CSS and JS files. By using the same setting in view.yml for site.js, main.css and other global, always-loaded assets you can arrive at a single minified file for all of the "standard" CSS and JS, with any extra stuff loading only as needed.

Note that for backwards compatibility nothing changes "out of the box" for existing sites, you have to set this up in app.yml and view.yml. See my forthcoming commit in the sandbox for a complete example.

If you are writing a plugin that uses two custom assets all of the time and a third custom asset only some of the time, you should use one asset group for the first two and another asset group for the remaining one so that the first two don't get needlessly recompiled and reloaded when the third is present.

Here's what view.yml looks like:

stylesheets:
  - main.css: { position: last, data-asset-group: global }
  - /apostropheExtraSlotsPlugin/css/aExtraSlots.css: { data-asset-group: global }

javascripts:
  - /apostropheBlogPlugin/js/aBlog.js: { data-asset-group: global }           # Required for blog plugin
  - /apostropheExtraSlotsPlugin/js/aMapSlot.js: { data-asset-group: global }
  - /apostropheCkEditorPlugin/js/ckeditor/ckeditor.js: { data-asset-group: global }
  - site.js: { position: last, data-asset-group: global }

And app.yml:

all:
  a:
    asset_group: global

Once you do this things will get a whole lot more efficient for cases where you have custom assets to be loaded only some of the time. If your only custom assets are already in view.yml then it won't make a difference since you are already avoiding any duplicate minifying and downloading of the same stuff.

We're not quite done here though. To get a really big win we need to be able to load things on demand for AJAX actions. Right now AJAX actions can't benefit from use_javascripts etc. in Apostrophe. That needs fixing



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3607 c89fab08-2f9b-42bf-8176-bd3ad6bee547