#1068 - History Button is broken for Singleton slots after the first use FIXED I re-factored how the history button gets its smarts. Instead of having a function call for every. single. history. button. There's a single function call in globalJavascripts that is executed if you're logged in that sets up every history button in one fell swoop.
The bug was relating to the fact that singleton slots have strange area/slot controls that were never fixed at the PHP level and there was a race condition with moving the control to the correct spot and applying the behaviors to those controls.

This was all avoided rather awesomely with the jquery .delegate() method.

The history button behavior now lives on the area itself and is delegated to the anchor.a-history-btn.

Awesome!



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3827 c89fab08-2f9b-42bf-8176-bd3ad6bee547