Fixed #1016
I brought the delete feature of the reorganize tree back using extremely simple markup and there are no issues affecting drag and drop etc. this time.

I'm using a cite element rather than any element that is already heavily styled by jstree (like a or span). And I'm using utterly simple inline markup. That made all the scary stuff go away without the need to aggressively re-style the tree and the breakage that resulted last time we tried to do that.

Although it doesn't look 100% consistent with our other stuff it works perfectly and we have clients, specifically FM, who simply can't fix certain problems on their own without it. 

If there is a desire to make this look like our other delete buttons I would strongly recommend *not* trying to float everything etc. as that would lead us to the same pain we experienced last time - please don't reopen that can of worms. Some kind of background image on the inline cite element might be okay. If you mess with it, test the heck out of it - especially the ability to get the "+" sign to appear and drop a page into another page, which is currently easy. 

Or... just leave the little red X alone in this one place, it works and we'll nail this in a whole new way in Apostrophe 2. That is my strong preference.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4026 c89fab08-2f9b-42bf-8176-bd3ad6bee547