20% speed improvement, especially when logged in, courtesy of Jeremy Kauffman. The basic issue: in many places we were writing...
url_for('a/foo')

This is very slow because Symfony has to check all the routes for the best match for that module. By using named routes every time we speed things up a lot.

However, Apostrophe doesn't have a name for every route, nor is that necessarily desirable for AJAX backend actions. So instead Jeremy added an a_url helper which constructs a URL for the requested module and action. By default, this helper expects that there is a route called 'default' that takes :module and :action from the URL, which is true in our sandbox and the great majority of sites. If it is not true for your sites, add a route called 'default' or set app_a_default_route to an alternate default route name.

This may cause some one-time inconvenience for a few existing sites that have no 'default' route, but the performance win is too great to ignore.

For your own code you should use named routes to start with, avoiding a_url or the foo/bar syntax. The consistent use of named routes has been a Symfony 1.x best practice for a while now for performance reasons.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4048 c89fab08-2f9b-42bf-8176-bd3ad6bee547