The new overrideLinks option to apostrophe.linkToRemote() can be used to inject an admin generator module or other really basic Symfony module that is not AJAX-aware into an AJAX container, such as a div.
All links, form submissions and delete buttons are rewritten on the fly to update the AJAX container. An event is notified on the container after these rewrites and event overrides are done for each update so that you have an opportunity to perform further overrides.

For this to work with an admin generator module you must shut off the layout for that module in view.yml, as the admin generator's "redirect to GET after POST" behavior breaks Symfony's ability to detect AJAX automatically. Implementing this for the delete buttons and their old-skool javascript onclick handlers was particularly gnarly, but the end result is pretty darn handy for a quick AJAX upgrade.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4177 c89fab08-2f9b-42bf-8176-bd3ad6bee547