Fixed #1197: aDate can be overridden without copying and pasting (extend BaseaDate)
Fixed #1200: you can write your own slot component if you really, really want to. We always intended to allow it, but it didn't actually work. Ironically I had to get rid of the executeSlot() convenience method in the base class to allow this to work because it causes false positives from componentExists()



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4415 c89fab08-2f9b-42bf-8176-bd3ad6bee547