#701 wide-but-not-tall logos unfairly punished by the use of netpbm
Dana, since the original is wider than 1024 pixels, we're falling back to using netpbm which can't preserve alpha channels, so you wind up with that nasty default to black background. 

Our restriction has been that if an image is bigger than 1024 pixels wide OR 768 pixels tall, we use netpbm instead of gd because of the very high memory requirements of loading big images in gd.

However this image isn't tall, it's just wide, so it's being unfairly punished.

So I improved the code to say what I really meant: if the image takes more than 4MB of RAM with gd (more than approximately one million pixels total), we use netpbm. Otherwise you get your transparency preserved. Images that happen to be wide but not tall are not "punished." This should be a big win for logos in general.

Now your logo is happy in dev:

[[Image(http://dl.dropbox.com/u/25049152/Screenshots/fm.png)]]

Everybody get chocolate medal.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4552 c89fab08-2f9b-42bf-8176-bd3ad6bee547