The new app_a_area_cache_enabled flag, when true, enables caching of individual content areas on the page for 5 minutes for logged-out users. This is a big win for sites that use expensive slots like the blog slot in sidebars that appear on many different pages but present the same content on each one. Especially when Google comes crawling. However  see app_a_page_cached_enabled, which is better for caching an entire page. These options are often enabled together.
Both singletons inserted with a_slot and full areas inserted with a_area are cached.

This feature is suitable for use with all of our standard slots, none of which pay attention to anything other than the options passed to them (and the current contents of the database, of course, so it may take 5 minutes for a logged out user to see changes made by an editor). If you have custom slots that look at URL parameters or other externalities this feature may not be for you. However note that it's OK to have links to ajax actions that care about such things from your normal view as long as the initial state of the slot is not dependent on URL parameters.

This feature has no impact on sites that do not explicitly enable the flag. We do not enable it on our sandbox.
 


git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4563 c89fab08-2f9b-42bf-8176-bd3ad6bee547